requirements.txt
flask>=3.0
python-dotenv>=1.0
requests>=2.31
dashscope>=1.25.11
flask-cors>=4.0
